Consolidate reduction gtests into a single test_reduction.cpp
Fold the original 16 parametrized sum/max/min cases into the
branch-keyed suite as six named tests (SumReduceW/H/BothDims,
MinMaxReduceW/H/BothDims), preserving their coverage exactly: tall
unaligned + aligned shapes per dim, positive and negative data ramps,
and both the single-tile HW single-core factory and the multi-tile
two-step path. test_reduction_branches.cpp is renamed over
test_reduction.cpp; one file, one style, 49 tests.

Note: the old SumTensor*/MinMaxTensor* gtest IDs are replaced by
ReductionBranches.* names.

Measured on BH p100a: 49/49 pass in 74 s (cold JIT for the new cells).
